Moved Wind class to Wind.py
